feat: add AXI master port
